requirements.txt
langgraph>=0.2.28
langchain>=0.3.1
langchain-openai>=0.2.1
tavily-python>=0.5.0
notion-client>=2.2.1
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.29.0
plotly>=5.18.0
pyperclip>=1.8.2
